+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Disable AdministrativeMonitor jenkins.security.ResourceDomainRecommendation
 
 ## [v2.414.3-1] - 2023-10-23
 ### Changed